[ecore] sqlDecoder uses zombie as sqlite lib

+++ ecore/sqlcodec.go
@@ -19,7 +19,7 @@ const (
 type SQLCodec struct {
 }
 
-const sqlCodecVersion int = 1
+const sqlCodecVersion int64 = 1
 
 func sqlTmpDB(prefix string) (string, error) {
 	try := 0
